Revert "refactor(print): remove the print layout from code, use template func…"
